--- src/components/action/button/index.ts
@@ -1,3 +1,0 @@
-export * from './button.component';
-export * from './button.interface';
-export * from './button.style';
